+++ src/slurm.rs
@@ -5,6 +5,7 @@ pub mod local {
 
     use crate::{Entity, Modifiable};
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn add_slurm_user(entity: &Entity, sacctmgr_path: &str) {
         let output = Command::new(sacctmgr_path)
             .arg("add")
@@ -37,10 +38,13 @@ pub mod local {
         }
 
         debug!("Modifying user qos");
-        modify_qos(entity, sacctmgr_path, true);
+        // Note: The order of execution is important here!
+        // Slurm expects the user to have QOS, before it can set the default QOS
         modify_qos(entity, sacctmgr_path, false);
+        modify_qos(entity, sacctmgr_path, true);
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn delete_slurm_user(user: &str, sacctmgr_path: &str) {
         let output = Command::new(sacctmgr_path)
             .arg("delete")
@@ -72,6 +76,7 @@ pub mod local {
         }
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn modify_slurm_user(modifiable: &Modifiable, sacctmgr_path: &str) {
         debug!("Start modifying user qos");
         match &modifiable.default_qos {
@@ -108,19 +113,21 @@ pub mod local {
         }
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn list_users(sacctmgr_path: &str) {
         let output = Command::new(sacctmgr_path)
-            .arg("list")
-            .arg("users")
-            .arg("format=User%30,DefaultAccount,Admin%15")
+            .arg("show")
+            .arg("assoc")
+            .arg("format=User%30,Account,DefaultQOS,QOS%80")
             .output()
             .expect(
                 "Unable to execute sacctmgr command. Is the path specified in your config correct?",
             );
-
+        // sacctmgr show assoc format=cluster,account,qos
         println!("{}", String::from_utf8_lossy(&output.stdout));
     }
 
+    /// TODO: Bubble up error instead of just logging it
     fn modify_qos(entity: &Entity, sacctmgr_path: &str, default_qos: bool) {
         let mut qos_str: String = "defaultQos=".to_owned();
         if default_qos {
@@ -129,7 +136,7 @@ pub mod local {
             let qos_joined = entity.qos.join(",");
             qos_str = format!("qos={}", qos_joined);
         }
-
+        debug!("Attempting to modify user with QOS: {qos_str}");
         let output = Command::new(sacctmgr_path)
             .arg("modify")
             .arg("user")
@@ -168,18 +175,18 @@ pub mod remote {
     use ssh2::Session;
 
     use crate::config::config::MgmtConfig;
-    use crate::{util::io_util::user_input, Entity, Modifiable};
+    use crate::ssh::SshCredential;
+    use crate::{Entity, Modifiable};
 
-    pub fn add_slurm_user(entity: &Entity, config: &MgmtConfig) {
-        let (ssh_username, ssh_password) = ask_credentials(&config.default_ssh_user);
-
+    /// TODO: Bubble up error instead of just logging it
+    pub fn add_slurm_user(entity: &Entity, config: &MgmtConfig, credentials: &SshCredential) {
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
         let tcp = TcpStream::connect(format!("{}:22", config.head_node)).unwrap();
         let mut sess = Session::new().unwrap();
 
         sess.handshake(&tcp).unwrap();
-        sess.userauth_password(&ssh_username, &ssh_password)
+        sess.userauth_password(credentials.username(), credentials.password())
             .unwrap();
 
         dbg!();
@@ -187,7 +194,7 @@ pub mod remote {
             "{} add user {} Account={} --immediate",
             config.sacctmgr_path, entity.username, entity.group
         );
-        let exit_code = run_command(&sess, &cmd);
+        let exit_code = run_remote_command(&sess, &cmd);
 
         match exit_code {
             0 => info!(
@@ -199,22 +206,27 @@ pub mod remote {
                 cmd
             ),
         };
+
+        debug!("Modifying user qos");
+        // Note: The order of execution is important here!
+        // Slurm expects the user to have QOS, before it can set the default QOS
+        modify_qos(&entity, config, &sess, false);
+        modify_qos(&entity, config, &sess, true);
     }
 
-    pub fn delete_slurm_user(user: &str, config: &MgmtConfig) {
-        let (ssh_username, ssh_password) = ask_credentials(&config.default_ssh_user);
-
+    /// TODO: Bubble up error instead of just logging it
+    pub fn delete_slurm_user(user: &str, config: &MgmtConfig, credentials: &SshCredential) {
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
         let tcp = TcpStream::connect(format!("{}:22", config.head_node)).unwrap();
         let mut sess = Session::new().unwrap();
 
         sess.handshake(&tcp).unwrap();
-        sess.userauth_password(&ssh_username, &ssh_password)
+        sess.userauth_password(credentials.username(), credentials.password())
             .unwrap();
 
         let cmd = format!("{} delete user {} --immediate", config.sacctmgr_path, user);
-        let exit_code = run_command(&sess, &cmd);
+        let exit_code = run_remote_command(&sess, &cmd);
 
         match exit_code {
             0 => info!("Successfully deleted Slurm user {}.", user),
@@ -225,16 +237,19 @@ pub mod remote {
         };
     }
 
-    pub fn modify_slurm_user(modifiable: &Modifiable, config: &MgmtConfig) {
-        let (ssh_username, ssh_password) = ask_credentials(&config.default_ssh_user);
-
+    /// TODO: Bubble up error instead of just logging it
+    pub fn modify_slurm_user(
+        modifiable: &Modifiable,
+        config: &MgmtConfig,
+        credentials: &SshCredential,
+    ) {
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
         let tcp = TcpStream::connect(format!("{}:22", config.head_node)).unwrap();
         let mut sess = Session::new().unwrap();
 
         sess.handshake(&tcp).unwrap();
-        sess.userauth_password(&ssh_username, &ssh_password)
+        sess.userauth_password(credentials.username(), credentials.password())
             .unwrap();
 
         debug!("Start modifying user default qos");
@@ -245,6 +260,7 @@ pub mod remote {
                     default_qos: m.to_string(),
                     ..Default::default()
                 };
+                debug!("New defaultQOS will be {}", entity.default_qos);
                 modify_qos(&entity, config, &sess, true);
             }
             None => info!(
@@ -260,6 +276,7 @@ pub mod remote {
                 qos: modifiable.qos.clone(),
                 ..Default::default()
             };
+            debug!("New QOS will be {:?}", entity.qos);
             modify_qos(&entity, config, &sess, false);
         } else {
             info!(
@@ -269,9 +286,9 @@ pub mod remote {
         }
     }
 
-    pub fn list_users(config: &MgmtConfig) {
-        let (ssh_username, ssh_password) = ask_credentials(&config.default_ssh_user);
-        let cmd = "sacctmgr list users";
+    /// TODO: Bubble up error instead of just logging it
+    pub fn list_users(config: &MgmtConfig, credentials: &SshCredential) {
+        let cmd = "sacctmgr show assoc format=User%30,Account,DefaultQOS,QOS%80";
 
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
@@ -279,7 +296,7 @@ pub mod remote {
         let mut sess = Session::new().unwrap();
 
         sess.handshake(&tcp).unwrap();
-        sess.userauth_password(&ssh_username, &ssh_password)
+        sess.userauth_password(credentials.username(), credentials.password())
             .unwrap();
 
         let mut channel = sess.channel_session().unwrap();
@@ -290,16 +307,7 @@ pub mod remote {
         println!("{}", s);
     }
 
-    fn ask_credentials(default_user: &str) -> (String, String) {
-        println!("Enter your SSH username (defaults to {}):", default_user);
-        let mut username = user_input();
-        if username.is_empty() {
-            username = default_user.to_string();
-        }
-        let password = rpassword::prompt_password("Enter your SSH password: ").unwrap();
-        (username, password)
-    }
-
+    /// TODO: Bubble up error instead of just logging it
     fn modify_qos(entity: &Entity, config: &MgmtConfig, sess: &Session, default_qos: bool) {
         let mut qos_str: String = "defaultQos=".to_owned();
         if default_qos {
@@ -308,12 +316,13 @@ pub mod remote {
             let qos_joined = entity.qos.join(",");
             qos_str = format!("qos={}", qos_joined);
         }
+        debug!("Attempting to modify user QOS with {qos_str}");
 
         let cmd = format!(
             "{} modify user {} set {} --immediate",
             config.sacctmgr_path, entity.username, qos_str
         );
-        let exit_code = run_command(sess, &cmd);
+        let exit_code = run_remote_command(sess, &cmd);
 
         match exit_code {
             0 => info!(
@@ -327,7 +336,11 @@ pub mod remote {
         };
     }
 
-    fn run_command(sess: &Session, cmd: &str) -> i32 {
+    /// Executes given command `cmd` on remote machine over ssh
+    ///
+    /// TODO: move checking if for exit code inside here and return result instead of error code
+    /// directly
+    fn run_remote_command(sess: &Session, cmd: &str) -> i32 {
         debug!("Running command {}", cmd);
 
         let mut channel = sess.channel_session().unwrap();